refactor(crm): extract read data access
P5-5: move the eight CRM read functions and six read-model interfaces out of
lib/crm.ts (2778 → 1878 lines) into lib/crm-reads.ts. Behavior-preserving —
every SQL string, binding, mapping, null fallback, sort order, limit, and return
shape is moved verbatim.

Moved to lib/crm-reads.ts:
- Functions: listBatches, getDashboardSummary, getBatchDetail, listContacts,
  listCompanies, listReviewTasks, listDrafts, getContactDetail.
- Interfaces: DashboardSummary, BatchListItem, BatchDetail, ContactDetail,
  ReviewTaskItem, DraftListItem.
- Imports only: getCrmDb, parseJsonValue, types from crm-types.

BatchImageRole moved from a private crm.ts declaration to lib/crm-types.ts
(exported); imported by both crm.ts (insertImage) and crm-reads.ts (image rows).

lib/crm.ts remains the public facade: imports and re-exports all eight functions
and six types. processApprovedBatch resolves the imported getBatchDetail binding
(not a detached export-from). No application caller changed — all still import
from @/lib/crm. Six now-unused type imports cleaned from crm.ts.

Verified: npm test (453 pass / 0 fail, +1 facade test), npm run lint (clean),
npm run build:cf (complete), git diff --check (clean). Audits: 0 of the 8
functions remain in crm.ts; 0 app callers import crm-reads directly; no import
cycle (crm-reads → cloudflare-runtime/crm-json/crm-types only).
